requirements-tradingview.txt
google-api-python-client
google-auth
google-auth-oauthlib
ccxt
pandas
pandas_ta
requests
